attendance_form.php works for last enrtry only.
attendance_form.php draws data from the swimmers table in the database,
and on submit, inserts the last entry of the list into the attendance
table with the correct swimmer ID. Big step forward for me! Now to just
get it to do all the rows, not just the last one.